Refactor logic for running cmd from TUI
Previously, the command was executed directly from the TUI, but
this led to issues where the program would sometimes crash for
some commands. This was changed to the command being ran after
the TUI quits. This fixed the issue because the TUI relinquishes
control of the terminal when it quits, which allows the command
to take over Stdin, Stdout, Stderr without issue.